+++ src/OkukusPages/Okukus.js
@@ -15,7 +15,7 @@ import Navigation from "./Nav/Nav";
 import Body from "./Body/Body";
 import Search from "./Search/Content";
 import Product from "./Product/Product";
-import OrderItem from "./Order/newOrder";
+import OrderItem from "./Order/Order";
 import Confirm from "./Confirmation/Confirm";
 import Cart from "./Cart/Cart";
 import Checkout from "./Checkout/Checkout";
